+++ src/events/FileEvent.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace luya\admin\events;
+
+use yii\base\Event;
+
+class FileEvent extends Event
+{
+    /**
+     * @var \luya\admin\models\StorageFile
+     */
+    public $file;
+}